@@ -1,7 +1,7 @@
 ext {
 
-    def kotlin_version = "1.5.32"
-    def androidx_lifecycle = "2.2.0"
+    def kotlin_version = "1.8.10"
+    def androidx_lifecycle = "2.5.1"
     def android_arch_lifecycle_version = "1.1.1"
     def com_square_up_retrofit2_version = "2.4.0"
     def room_version = "2.2.3"
@@ -15,7 +15,8 @@ ext {
     def app_compat_version = "1.4.0"
     def constraint_layout_version = "2.1.2"
     def espresso_version = "3.4.0"
-    def hilt_version = "2.41"
+    def hilt_version = "2.45"
+    def androidx_test_junit_version = "1.1.5"
 
     appDependencies = [
             "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version",
@@ -25,6 +26,7 @@ ext {
             "androidx.constraintlayout:constraintlayout:$constraint_layout_version",
             "androidx.lifecycle:lifecycle-viewmodel-ktx:$androidx_lifecycle",
             "androidx.lifecycle:lifecycle-livedata-ktx:$androidx_lifecycle",
+            "androidx.navigation:navigation-fragment-ktx:$androidx_lifecycle",
             "io.reactivex.rxjava2:rxjava:$rx_java_version",
             'io.reactivex.rxjava2:rxandroid:2.1.1',
             "android.arch.lifecycle:extensions:$android_arch_lifecycle_version",
@@ -37,7 +39,7 @@ ext {
             "com.google.android.material:material:$material_version",
             "com.github.bumptech.glide:glide:$glide_version",
             "com.github.bumptech.glide:annotations:$glide_version",
-            "com.google.dagger:hilt-android:$hilt_version"
+            "com.google.dagger:hilt-android:$hilt_version",
     ]
 
     appKaptDependencies = [
@@ -91,7 +93,7 @@ ext {
     ]
 
     appAndroidTestImplementation = [
-            'androidx.test.ext:junit:1.1.3',
+            "androidx.test.ext:junit:$androidx_test_junit_version",
             "androidx.test.espresso:espresso-core:$espresso_version",
             "androidx.test.espresso:espresso-contrib:$espresso_version",
             'org.junit.jupiter:junit-jupiter-api:5.8.2'
@@ -104,8 +106,8 @@ ext {
     ]
 
     dataAndroidTestImplementation = [
-            'androidx.test.ext:junit:1.1.1',
-            'androidx.test.espresso:espresso-core:3.2.0',
+            "androidx.test.ext:junit:$androidx_test_junit_version",
+            "androidx.room:room-testing:$room_version",
             "junit:junit:$junit_version"
     ]
 }
