@@ -1,13 +1,14 @@
 package com.haomins.reader.adapters
 
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
+import android.widget.ImageView
+import android.widget.TextView
+import androidx.constraintlayout.widget.ConstraintLayout
 import androidx.recyclerview.widget.RecyclerView
 import com.haomins.domain.model.entities.ArticleEntity
-import com.haomins.reader.R
+import com.haomins.reader.databinding.ArticleTitleRecyclerViewItemBinding
 import com.haomins.reader.utils.GlideUtils
-import kotlinx.android.synthetic.main.article_title_recycler_view_item.view.*
 
 class ArticleTitleListAdapter(
     private val articleTitleListUiItems: List<ArticleEntity>,
@@ -24,15 +25,16 @@ class ArticleTitleListAdapter(
 
     }
 
-    inner class CustomViewHolder(val viewHolder: View) : RecyclerView.ViewHolder(viewHolder)
-
     override fun onCreateViewHolder(
         parent: ViewGroup,
         viewType: Int
     ): ArticleTitleListAdapter.CustomViewHolder {
-        val readerArticleItemView = LayoutInflater.from(parent.context)
-            .inflate(R.layout.article_title_recycler_view_item, parent, false)
-        return CustomViewHolder(readerArticleItemView)
+        val itemBinding = ArticleTitleRecyclerViewItemBinding.inflate(
+            LayoutInflater.from(parent.context),
+            parent,
+            false
+        )
+        return CustomViewHolder(itemBinding)
     }
 
     override fun getItemCount(): Int {
@@ -40,23 +42,37 @@ class ArticleTitleListAdapter(
     }
 
     override fun onBindViewHolder(holder: ArticleTitleListAdapter.CustomViewHolder, position: Int) {
-        with(holder.viewHolder) {
-            articleTitleListUiItems[position].let {
-                article_title.text = it.itemTitle
-                article_posted_time.text = it.howLongAgo
-                onArticleClicked(it.itemId)
-                glideUtils.loadPreviewImage(
-                    imageView = article_preview_image,
-                    it.previewImageUrl
-                )
-            }
-        }
+        holder.bind(articleTitleListUiItems[position], position)
         articleTitleListOnClickListener.onLoadMoreArticlesBasedOnPosition(position)
     }
 
-    private fun View.onArticleClicked(articleItemId: String) {
-        setOnClickListener {
-            articleTitleListOnClickListener.onArticleClicked(articleItemId)
+    inner class CustomViewHolder(val binding: ArticleTitleRecyclerViewItemBinding) :
+        RecyclerView.ViewHolder(binding.root) {
+
+        private val articleTitleTextView: TextView
+        private val articlePostedTimeTextView: TextView
+        private val articlePreviewImageView: ImageView
+        private val itemRootView: ConstraintLayout
+
+        init {
+            articleTitleTextView = binding.articleTitle
+            articlePostedTimeTextView = binding.articlePostedTime
+            articlePreviewImageView = binding.articlePreviewImage
+            itemRootView = binding.root
+        }
+
+        fun bind(articleEntity: ArticleEntity, position: Int) {
+            with(articleEntity) {
+                articleTitleTextView.text = itemTitle
+                articlePostedTimeTextView.text = howLongAgo
+                glideUtils.loadPreviewImage(
+                    articlePreviewImageView,
+                    previewImageUrl
+                )
+                itemRootView.setOnClickListener {
+                    articleTitleListOnClickListener.onArticleClicked(itemId)
+                }
+            }
         }
     }
 
