@@ -4,12 +4,11 @@ import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import com.haomins.reader.R
+import com.haomins.reader.databinding.ActivityArticleDetailBinding
 import com.haomins.reader.utils.slideInAnimation
 import com.haomins.reader.utils.slideOutAnimation
-import com.haomins.reader.view.fragments.ArticleDetailFragment
+import com.haomins.reader.view.fragments.articles.ArticleDetailFragment
 import dagger.hilt.android.AndroidEntryPoint
-import kotlinx.android.synthetic.main.activity_article_detail.*
 
 @AndroidEntryPoint
 class ArticleDetailActivity : AppCompatActivity() {
@@ -18,10 +17,12 @@ class ArticleDetailActivity : AppCompatActivity() {
         const val ARTICLE_ITEM_ID = "ARTICLE_ITEM_ID"
     }
 
+    private lateinit var binding: ActivityArticleDetailBinding
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        inflateView()
         slideInAnimation()
-        setContentView(R.layout.activity_article_detail)
         initViewPager()
     }
 
@@ -39,9 +40,14 @@ class ArticleDetailActivity : AppCompatActivity() {
                 this,
                 it
             )
-            article_detail_view_pager.adapter = adapter
-            article_detail_view_pager.setCurrentItem(currentPosition, false)
+            binding.articleDetailViewPager.adapter = adapter
+            binding.articleDetailViewPager.setCurrentItem(currentPosition, false)
         }
+    }
+
+    private fun inflateView() {
+        binding = ActivityArticleDetailBinding.inflate(layoutInflater)
+        setContentView(binding.root)
     }
 
     private inner class ArticleDetailFragmentAdapter(
