@@ -4,21 +4,24 @@ import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.FragmentTransaction
 import com.haomins.reader.R
+import com.haomins.reader.databinding.ActivitySettingsBinding
 import com.haomins.reader.utils.slideAnimation
 import com.haomins.reader.utils.slideInAnimation
 import com.haomins.reader.utils.slideOutAnimation
-import com.haomins.reader.view.fragments.AboutFragment
-import com.haomins.reader.view.fragments.DisclosureFragment
-import com.haomins.reader.view.fragments.SettingsFragment
+import com.haomins.reader.view.fragments.settings.AboutFragment
+import com.haomins.reader.view.fragments.settings.DisclosureFragment
+import com.haomins.reader.view.fragments.settings.SettingsFragment
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
 class SettingsActivity : AppCompatActivity() {
 
+    private lateinit var binding: ActivitySettingsBinding
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        inflateView()
         slideInAnimation()
-        setContentView(R.layout.activity_settings)
         showSettingsFragment()
     }
 
@@ -62,6 +65,11 @@ class SettingsActivity : AppCompatActivity() {
                 SettingsFragment.TAG
             )
             .commit()
+    }
+
+    private fun inflateView() {
+        binding = ActivitySettingsBinding.inflate(layoutInflater)
+        setContentView(binding.root)
     }
 
 }
