@@ -11,16 +11,17 @@ import androidx.fragment.app.viewModels
 import androidx.lifecycle.Observer
 import com.haomins.reader.BuildConfig
 import com.haomins.reader.R
+import com.haomins.reader.databinding.FragmentLoginBinding
 import com.haomins.reader.utils.showToast
 import com.haomins.reader.view.activities.MainActivity
 import com.haomins.reader.viewModels.LoginViewModel
 import dagger.hilt.android.AndroidEntryPoint
-import kotlinx.android.synthetic.main.fragment_login.*
 
 @AndroidEntryPoint
 class LoginFragment : Fragment() {
 
     private val loginViewModel by viewModels<LoginViewModel>()
+    private lateinit var binding: FragmentLoginBinding
 
     private val isUserLoggedInObserver by lazy {
         Observer<Boolean> {
@@ -35,8 +36,9 @@ class LoginFragment : Fragment() {
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? {
-        return inflater.inflate(R.layout.fragment_login, container, false)
+    ): View {
+        binding = FragmentLoginBinding.inflate(inflater, container, false)
+        return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -48,10 +50,12 @@ class LoginFragment : Fragment() {
     }
 
     private fun setOnClickListener() {
-        login_button.setOnClickListener { loginButtonOnClick() }
-        sign_up_button.setOnClickListener { signUpButtonOnClick() }
-        forgot_password.setOnClickListener { forgetPasswordButtonOnClick() }
-        news_app_disclosure.setOnClickListener { showDisclosureFragment() }
+        with(binding) {
+            loginButton.setOnClickListener { loginButtonOnClick() }
+            signUpButton.setOnClickListener { signUpButtonOnClick() }
+            forgotPassword.setOnClickListener { forgetPasswordButtonOnClick() }
+            newsAppDisclosure.setOnClickListener { showDisclosureFragment() }
+        }
     }
 
     private fun showDisclosureFragment() {
@@ -63,13 +67,15 @@ class LoginFragment : Fragment() {
     }
 
     private fun setLoginPasswordEditTextOnChangeListener() {
-        login_password_edit_text.addTextChangedListener {
-            login_button.isEnabled = !it.isNullOrEmpty()
+        with(binding) {
+            loginPasswordEditText.addTextChangedListener {
+                loginButton.isEnabled = !it.isNullOrEmpty()
+            }
         }
     }
 
     private fun initiateUI() {
-        login_app_version_text_view.text =
+        binding.loginAppVersionTextView.text =
             getString(R.string.version_description, BuildConfig.VERSION_NAME)
     }
 
@@ -90,11 +96,13 @@ class LoginFragment : Fragment() {
     }
 
     private fun loginButtonOnClick() {
-        loginViewModel.authorize(
-            userName = login_username_edit_text.text.toString(),
-            userPassword = login_password_edit_text.text.toString()
-        ) {
-            showToast(it)
+        with(binding) {
+            loginViewModel.authorize(
+                userName = loginUsernameEditText.text.toString(),
+                userPassword = loginPasswordEditText.text.toString()
+            ) {
+                showToast(it)
+            }
         }
     }
 
